P004_YJ_Crawler/requirements.txt
requests~=2.28.2
pandas~=1.5.3
xlrd~=2.0.1
PyQt5>=5.15.9
pycryptodome~=3.17